requirements.txt
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6-sip==13.10.0
requests
pycryptodome
